+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/process/package.scala
@@ -1,5 +1,5 @@
 package pl.touk.nussknacker.engine.api
 
 package object process {
-  @deprecated("Renamed to ComponentUsaCase", "1.3") type RunMode = ComponentUseCase
+  @deprecated("Renamed to ComponentUseCase", "1.3") type RunMode = ComponentUseCase
 }