+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/registrar/FlinkProcessRegistrar.scala
@@ -8,7 +8,6 @@ import org.apache.flink.streaming.api.scala.{DataStream, OutputTag, StreamExecut
 import org.apache.flink.streaming.api.windowing.time.Time
 import pl.touk.nussknacker.engine.InterpretationResult
 import pl.touk.nussknacker.engine.api._
-import pl.touk.nussknacker.engine.api.async.{DefaultAsyncInterpretationValue, DefaultAsyncInterpretationValueDeterminer}
 import pl.touk.nussknacker.engine.api.component.NodeComponentInfo
 import pl.touk.nussknacker.engine.api.context.{JoinContextTransformation, ValidationContext}
 import pl.touk.nussknacker.engine.compiledgraph.part._
@@ -19,13 +18,14 @@ import pl.touk.nussknacker.engine.flink.api.compat.ExplicitUidInOperatorsSupport
 import pl.touk.nussknacker.engine.flink.api.process._
 import pl.touk.nussknacker.engine.flink.api.typeinformation.TypeInformationDetection
 import pl.touk.nussknacker.engine.graph.EspProcess
-import pl.touk.nussknacker.engine.graph.node.BranchEndDefinition
+import pl.touk.nussknacker.engine.graph.node.{BranchEndDefinition, NodeData}
 import pl.touk.nussknacker.engine.process.compiler.{FlinkEngineRuntimeContextImpl, FlinkProcessCompiler, FlinkProcessCompilerData}
 import pl.touk.nussknacker.engine.process.typeinformation.TypeInformationDetectionUtils
 import pl.touk.nussknacker.engine.process.util.StateConfiguration.RocksDBStateBackendConfig
 import pl.touk.nussknacker.engine.process.{CheckpointConfig, ExecutionConfigPreparer, FlinkCompatibilityProvider}
 import pl.touk.nussknacker.engine.resultcollector.{ProductionServiceInvocationCollector, ResultCollector}
 import pl.touk.nussknacker.engine.splittedgraph.end.BranchEnd
+import pl.touk.nussknacker.engine.splittedgraph.splittednode
 import pl.touk.nussknacker.engine.splittedgraph.splittednode.EndingNode
 import pl.touk.nussknacker.engine.testmode.{SinkInvocationCollector, TestRunId, TestServiceInvocationCollector}
 import pl.touk.nussknacker.engine.util.loader.ScalaServiceLoader
@@ -42,6 +42,8 @@ import scala.language.implicitConversions
  */
 class FlinkProcessRegistrar(compileProcess: (EspProcess, ProcessVersion, DeploymentData, ResultCollector) => ClassLoader => FlinkProcessCompilerData,
                             streamExecutionEnvPreparer: StreamExecutionEnvPreparer) extends LazyLogging {
+
+  import FlinkProcessRegistrar._
 
   implicit def millisToTime(duration: Long): Time = Time.of(duration, TimeUnit.MILLISECONDS)
 
@@ -119,7 +121,7 @@ class FlinkProcessRegistrar(compileProcess: (EspProcess, ProcessVersion, Deploym
         .sourceStream(env, nodeContext(nodeComponentInfoFrom(part), Left(ValidationContext.empty)))
         .process(new SourceMetricsFunction(part.id))(contextTypeInformation)
 
-      val asyncAssigned = registerInterpretationPart(start, part, "interpretation")
+      val asyncAssigned = registerInterpretationPart(start, part, InterpretationName)
 
       registerNextParts(asyncAssigned, part)
     }
@@ -146,7 +148,7 @@ class FlinkProcessRegistrar(compileProcess: (EspProcess, ProcessVersion, Deploym
         .transform(inputs.mapValues(_._1), nodeContext(nodeComponentInfoFrom(joinPart), Right(inputs.mapValues(_._2))))
         .map(newContextFun)(typeInformationDetection.forContext(joinPart.validationContext))
 
-      val afterSplit = registerInterpretationPart(newStart, joinPart, "branchInterpretation")
+      val afterSplit = registerInterpretationPart(newStart, joinPart, BranchInterpretationName)
       registerNextParts(afterSplit, joinPart)
     }
 
@@ -198,7 +200,7 @@ class FlinkProcessRegistrar(compileProcess: (EspProcess, ProcessVersion, Deploym
             .addSink(new CollectingSinkFunction[AnyRef](compiledProcessWithDeps, collectingSink, part.id))
       }
 
-      withSinkAdded.name(s"${metaData.id}-${part.id}-sink")
+      withSinkAdded.name(operatorName(metaData, part.node, "sink"))
       Map()
     }
 
@@ -223,7 +225,7 @@ class FlinkProcessRegistrar(compileProcess: (EspProcess, ProcessVersion, Deploym
             case None => ir.context
           }
           val newStart = transformed.map(newContextFun)(typeInformationDetection.forContext(part.validationContext))
-          val afterSplit = registerInterpretationPart(newStart, part, "customNodeInterpretation")
+          val afterSplit = registerInterpretationPart(newStart, part, CustomNodeInterpretationName)
           registerNextParts(afterSplit, part)
       }
     }
@@ -241,11 +243,9 @@ class FlinkProcessRegistrar(compileProcess: (EspProcess, ProcessVersion, Deploym
       val metaData = processWithDeps.metaData
       val streamMetaData = MetaDataExtractor.extractTypeSpecificDataOrFail[StreamMetaData](metaData)
 
-      val useIOMonad = globalParameters.flatMap(_.configParameters).flatMap(_.useIOMonadInInterpreter).getOrElse(true)
-      //TODO: we should detect automatically that Interpretation has no async enrichers and invoke sync function then, as async comes with
-      //performance penalty...
-      val defaultAsync: DefaultAsyncInterpretationValue = DefaultAsyncInterpretationValueDeterminer.determine(asyncExecutionContextPreparer)
-      val shouldUseAsyncInterpretation = streamMetaData.useAsyncInterpretation.getOrElse(defaultAsync.value)
+      val configParameters = globalParameters.flatMap(_.configParameters)
+      val useIOMonad = configParameters.flatMap(_.useIOMonadInInterpreter).getOrElse(true)
+      val shouldUseAsyncInterpretation = AsyncInterpretationDeterminer(configParameters, asyncExecutionContextPreparer).determine(node, streamMetaData)
       (if (shouldUseAsyncInterpretation) {
         val asyncFunction = new AsyncInterpretationFunction(compiledProcessWithDeps, node, validationContext, asyncExecutionContextPreparer, useIOMonad)
         ExplicitUidInOperatorsSupport.setUidIfNeed(ExplicitUidInOperatorsSupport.defaultExplicitUidInStatefulOperators(globalParameters), node.id + "-$async")(
@@ -254,9 +254,10 @@ class FlinkProcessRegistrar(compileProcess: (EspProcess, ProcessVersion, Deploym
       } else {
         val ti = InterpretationResultTypeInformation.create(typeInformationDetection, outputContexts)
         stream.flatMap(new SyncInterpretationFunction(compiledProcessWithDeps, node, validationContext, useIOMonad))(ti)
-      }).name(s"${metaData.id}-${node.id}-$name")
+      }).name(interpretationOperatorName(metaData, node, name, shouldUseAsyncInterpretation))
         .process(new SplitFunction(outputContexts, typeInformationDetection))(org.apache.flink.streaming.api.scala.createTypeInformation[Unit])
     }
+
   }
 
   private def nodeComponentInfoFrom(processPart: ProcessPart): NodeComponentInfo = {
@@ -267,6 +268,9 @@ class FlinkProcessRegistrar(compileProcess: (EspProcess, ProcessVersion, Deploym
 object FlinkProcessRegistrar {
 
   private[registrar] final val EndId = "$end"
+  final val InterpretationName = "interpretation"
+  final val CustomNodeInterpretationName = "customNodeInterpretation"
+  final val BranchInterpretationName = "branchInterpretation"
 
   import net.ceedubs.ficus.Ficus._
   import net.ceedubs.ficus.readers.ArbitraryTypeReader._
@@ -282,6 +286,26 @@ object FlinkProcessRegistrar {
         .headOption.map(_.createExecutionEnvPreparer(config, prepareExecutionConfig, compiler.diskStateBackendSupport))
         .getOrElse(new DefaultStreamExecutionEnvPreparer(checkpointConfig, rocksDBStateBackendConfig, prepareExecutionConfig))
     new FlinkProcessRegistrar(compiler.compileProcess, defaultStreamExecutionEnvPreparer)
+  }
+
+  private[registrar] def operatorName(metaData: MetaData,
+                                      splittedNode: splittednode.SplittedNode[NodeData],
+                                      operation: String) = {
+    s"${metaData.id}-${splittedNode.id}-$operation"
+  }
+
+  private[registrar] def interpretationOperatorName(metaData: MetaData,
+                                                    splittedNode: splittednode.SplittedNode[NodeData],
+                                                    interpretationName: String,
+                                                    shouldUseAsyncInterpretation: Boolean): String = {
+    interpretationOperatorName(metaData.id, splittedNode.id, interpretationName, shouldUseAsyncInterpretation)
+  }
+
+  private[registrar] def interpretationOperatorName(scenarioId: String,
+                                                    nodeId: String,
+                                                    interpretationName: String,
+                                                    shouldUseAsyncInterpretation: Boolean) = {
+    s"$scenarioId-$nodeId-$interpretationName${if (shouldUseAsyncInterpretation) "Async" else "Sync"}"
   }
 
 }