+++ engine/flink/executor/src/test/scala/pl/touk/nussknacker/engine/process/runner/FlinkTestMainSpec.scala
@@ -399,8 +399,8 @@ class FlinkTestMainSpec extends FunSuite with Matchers with Inside with BeforeAn
     val process = ScenarioBuilder
       .streaming("proc")
       .source("start", "input")
-      .enricher("componentUseCaseService", "componentUseCaseService", "returningComponentUsaCaseService")
-      .customNode("componentUseCaseCustomNode", "componentUseCaseCustomNode", "transformerAddingComponentUsaCase")
+      .enricher("componentUseCaseService", "componentUseCaseService", "returningComponentUseCaseService")
+      .customNode("componentUseCaseCustomNode", "componentUseCaseCustomNode", "transformerAddingComponentUseCase")
       .emptySink("out", "valueMonitor", "value" -> "{#componentUseCaseService, #componentUseCaseCustomNode}")
 
     val results = runFlinkTest(process, TestData.newLineSeparated("0|1|2|3|4|5|6"))